// src/codegen/google/protobuf/timestamp.d.ts
export interface Timestamp {
    /**
     * Represents seconds of UTC time since Unix epoch
     * 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
     * 9999-12-31T23:59:59Z inclusive.
     */
    seconds: bigint;
    /**
     * Non-negative fractions of a second at nanosecond resolution. Negative
     * second values with fractions must still have non-negative nanos values
     * that count forward in time. Must be from 0 to 999,999,999
     * inclusive.
     */
    nanos: number;
}
// ...
export declare const Timestamp: {
    typeUrl: string;
    encode(message: Timestamp, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): Timestamp;
    fromPartial(object: Partial<Timestamp>): Timestamp;
    fromAmino(object: TimestampAmino): Timestamp;
    toAmino(message: Timestamp): TimestampAmino;
    fromAminoMsg(object: TimestampAminoMsg): Timestamp;
    fromProtoMsg(message: TimestampProtoMsg): Timestamp;
    toProto(message: Timestamp): Uint8Array;
    toProtoMsg(message: Timestamp): TimestampProtoMsg;
};